Changed OverlapController signature, fixed duplicate code
